Refactored index.ts
