+++ src/Console/ANSIColourEscapeSequence.ts
